requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.4.1 # https://github.com/andymccurdy/redis-py

# Django
# ------------------------------------------------------------------------------
django==2.0.13  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-ajax-selects==1.7.1  # https://pypi.org/project/django-ajax-selects/1.7.1/
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms

# PDF
xhtml2pdf==0.2.3
django-xhtml2pdf==0.0.4
